+++ app/src/main/java/com/selasarimaji/perpus/viewmodel/EditKidVM.kt
@@ -3,8 +3,8 @@ package com.selasarimaji.perpus.viewmodel
 import android.util.Log
 import com.google.firebase.firestore.DocumentChange
 import com.selasarimaji.perpus.model.DataModel
-import com.selasarimaji.perpus.repository.firebase.BaseRepo
-import com.selasarimaji.perpus.repository.firebase.KidRepo
+import com.selasarimaji.perpus.repository.firestore.BaseRepo
+import com.selasarimaji.perpus.repository.firestore.KidRepo
 
 class EditKidVM : BaseContentVM<DataModel.Kid>() {
     private val repoVal by lazy {
@@ -17,18 +17,21 @@ class EditKidVM : BaseContentVM<DataModel.Kid>() {
         get() = EditKidVM::class.java.name
 
     override fun loadInitial() {
-        repo.loadRange(0, 10){ querySnapshot, exception ->
-            if (exception != null) {
-                Log.w(TAG, "listen:error", exception)
-            }else {
-                querySnapshot!!.documentChanges.map {
-                    when(it.type){
-                        DocumentChange.Type.ADDED -> repo.addLocalItem(DataModel.Kid.turnDocumentToObject(it.document))
-                        DocumentChange.Type.MODIFIED -> repo.editLocalItem(DataModel.Kid.turnDocumentToObject(it.document))
-                        DocumentChange.Type.REMOVED -> repo.deleteLocalItem(DataModel.Kid.turnDocumentToObject(it.document))
+        if (isInitialLoaded.value == null || !isInitialLoaded.value!!){
+            repo.loadRange(0, 10) { querySnapshot, exception ->
+                if (exception != null) {
+                    Log.w(TAG, "listen:error", exception)
+                } else {
+                    querySnapshot!!.documentChanges.map {
+                        when (it.type) {
+                            DocumentChange.Type.ADDED -> repo.addLocalItem(DataModel.Kid.turnDocumentToObject(it.document))
+                            DocumentChange.Type.MODIFIED -> repo.editLocalItem(DataModel.Kid.turnDocumentToObject(it.document))
+                            DocumentChange.Type.REMOVED -> repo.deleteLocalItem(DataModel.Kid.turnDocumentToObject(it.document))
+                        }
                     }
                 }
             }
+            isInitialLoaded.value = true
         }
     }
 }
